JZ6AL-22 : move app menu logic from app to WwrController
